more directionlaity fixes with accepting/denying friend requests

+++ flask-server/database.py
@@ -169,6 +169,12 @@ class Database():
                 session.delete(friend_request)
                 session.commit()
                 session.expire_all()
+            
+            friend_request = session.query(Requests).filter_by(user_id=friend_id, friend_id=user_id).first()
+            if friend_request:
+                session.delete(friend_request)
+                session.commit()
+                session.expire_all()
 
     def accept_friend_request(self, user_id, friend_id):
         with self.get_session() as session:
@@ -178,6 +184,16 @@ class Database():
                 new_friendship = Friend(
                     user_id=user_id,
                     friend_id=friend_id
+                )
+                session.add(new_friendship)
+                session.delete(friend_request)
+                session.commit()
+                session.expire_all()
+            friend_request = session.query(Requests).filter_by(user_id=friend_id, friend_id=user_id).first()
+            if friend_request:
+                new_friendship = Friend(
+                    user_id=friend_id,
+                    friend_id=user_id
                 )
                 session.add(new_friendship)
                 session.delete(friend_request)
